Refatorado servilo de documento
